[util][runtime] Tried to split package and functi
- #32
- tried to split package and function name from frame's return value
using string.LastIndex, but this does not work, because we need to
consider slash
